Fixed the bug in UpdateNew component, merged studentforms and Form component, added new Functionality inside Navbar
