+++ src/context/FormContext.js
@@ -12,17 +12,14 @@ const defaultData = {
 	phoneNumber: "",
 	cgpa: "",
 	address: "",
-    filled: null,
 };
 export const Datacontext= createContext();
 export const DataProvider= (props)=>{
     const [formData, setformData] = useState(defaultData);
-    const [submit,setSubmit] = useState("Submit");
     const [flag, setFlag] = useState(false);
     return (
         <Datacontext.Provider value={{
             dataValues: [formData, setformData], 
-            dataSubmit: [submit, setSubmit],
             dataFlag: [flag, setFlag]}}>
             {props.children}
         </Datacontext.Provider>
